requirements.txt
streamlit
pandas
altair
Pillow
python-dotenv
google-genai